% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alignSummarySTAR.R
\name{alignSummarySTAR}
\alias{alignSummarySTAR}
\title{Extracts the alignment summary from STAR log files.}
\usage{
alignSummarySTAR(logFiles, logPath, pattern = "Log.final.out")
}
\arguments{
\item{logFiles}{A \code{vector} of log files. Any files which cannot be found will be ignored.
If \code{logPath} is not specified, these must contain path information as part of each filename}

\item{logPath}{The path to the directory where the log files are located.
If no files are specified, this will default to all files which match the pattern "Log.out"}

\item{pattern}{A text string found in the name of all required logFiles. Not used if the files are specified by name.}
}
\value{
a list with components \code{$reads, $splicing, $lengths, $indels} and \code{$performance}.
Each of thsi will be a \code{data.frame} with samples as rows and the key information as columns.
}
\description{
Extracts the basic summary for fastq files after running fastqc.
}
\details{
This will open the log files as output by the alignment tool STAR and load the information into your workspace.
The log files in question are the ones traditionally output with the suffix "Log.final.out"
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}

