% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bowtie.R
\name{buildBowtie}
\alias{buildBowtie}
\title{A wrapper for the bash shell tool bowtie.}
\usage{
buildBowtie(genome, ebwt, genDir, indDir, args = c(),
  exec = "/usr/bin/bowtie-build")
}
\arguments{
\item{genome}{The reference genome as a single fasta file.}

\item{ebwt}{The prefix to give the index files. Will default to the basename of the reference genome.}

\item{genDir}{The directory containing the reference genome. Only required if not specified as part of the genome filename.}

\item{args}{Any additional arguments/options to pass to the commad}

\item{exec}{the path to the executable. This may vary from system to system}

\item{inDir}{The directory to write the index files to. Will default to the same directory as the reference genome}
}
\value{
a \code{list} with components \code{$log}, \code{$ebwt}, \code{$version}, \code{$command} and \code{$time}
}
\description{
Builds a bowtie index, or aligns to the genome using bowtie.
}
\details{
\code{bowtieBuild} will build the index files required for aligning using bowtie, or tophat.
The list returned by this function will contain the \code{stdout} information in the component \code{$log}.
The \code{ebwt} prefix required by the subsequent functions will be in the component \code{$ebwt}.
The remaining list components contain the \code{system} command as executed, the version information and the date the index was built.

Be aware that this can be a lengthy process depending on the genome size.
Please be patient and check the output folder to see if files are still being written to.

A wrapper for bowtie is still under construction and hasn't been written yet
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}

