% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/countChastityReads.R
\name{countChastityReads}
\alias{countChastityReads}
\title{Find how many reads in a file pass or fail the Illumina Chastity Filter.}
\usage{
countChastityReads(inFile)
}
\arguments{
\item{inFile}{the name of the file to be queried. Must contain the complete path}
}
\value{
A \code{data.frame} as described above
}
\description{
Takes an individual fastq file and counts the reads with the Y/N Chastity flag set
}
\details{
Returns a data.frame with four components:

\code{Filename} the name of the fastq file being queried.

\code{Total} contains the total number of reads in the fastq file.

\code{Pass} contains the number of reads which pass the Illumina filter.

\code{Fail} contains the number of reads which fail the Illumina filter.
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}

