% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cutadaptPairedSummary.R
\name{cutadaptPairedSummary}
\alias{cutadaptPairedSummary}
\title{Extract Read Information From Cutadapt Log Files}
\usage{
cutadaptPairedSummary(x)
}
\arguments{
\item{x}{The output from \code{cutadaptPaired}}
}
\value{
A \code{data.frame}
}
\description{
Takes the output from \code{cutadaptPaired} and extracts key information
}
\details{
This will extract the read numbers, the numbers trimmed & the numbers discarded from each file in the paired reads.
If the log files are from a set of pairs, it will automatically detect this, and get the information for the entire set of paris
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}

