% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastqIlluminaFilter.R
\name{fastqIlluminaFilter}
\alias{fastqIlluminaFilter}
\title{Remove reads from a fastq file that fail the Illumina Chastity Filter}
\usage{
fastqIlluminaFilter(inFile, outDir, args = "-N -v",
  exec = "/usr/local/bin/fastq_illumina_filter")
}
\arguments{
\item{inFile}{the name of the file to be processed. Must contain the complete path}

\item{outDir}{the name of the directory to write the files to.}

\item{args}{any additional arguments to specify}

\item{exec}{the path to the executable}
}
\value{
The text from the logfile, from the command line tool \code{fastq_illumina_filter}
}
\description{
A wrapper to the tool \code{fastq_illumina_filter} which must be installed on the local machine.
}
\details{
Returns the log file from the process
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}

