% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fqcBasic.R
\name{fqcBasic}
\alias{extractFqcBasic}
\alias{fqcBasic}
\title{Extracts the basic summary from fastqc reports.}
\usage{
fqcBasic(fqNames, qcDir)

extractFqcBasic(fqName, qcDir)
}
\arguments{
\item{fqNames}{a vector of filenames to extract the totals for.}

\item{qcDir}{the directory to look for the FASTQC reports}

\item{fqName}{the filename to extract the totals for.
This should be the name of a single fastq file.}
}
\value{
a \code{data.frame} with the basic summary information as contained in FASTQC reports
}
\description{
Extracts the basic summary for fastq files after running fastqc.
}
\details{
Always use the function \code{fqcBasic} as this will extract the basic summary information for one or many
 fastqc reports. The function \code{extractFqcBasic} is the low-level function called by the function \code{fqcBasic}
 and is designed to extract information from only a single \code{FASTQC} report.
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}

