% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fqcSummary.R
\name{extractFqcSummary}
\alias{extractFqcSummary}
\alias{fqcSummary}
\alias{fqcSummaryPlot}
\title{Extract the PASS/FAIL information from a batch of FASTQC summaries}
\usage{
extractFqcSummary(fqName, qcDir)

fqcSummary(fqNames, qcDir)

fqcSummaryPlot(fqNames, qcDir)
}
\arguments{
\item{fqName}{the filename to extract the totals for.
This should be the name of a single fastq file.}

\item{qcDir}{the directory to look in for the FASTQC reports}

\item{fqNames}{the filenames to extract the totals for.}
}
\value{
the PASS/FAIL summary information from the FASTQC report(s) either as a
data.frame (\code{fqcSummary}) or as a plot (\code{fqcSummaryPlot})
}
\description{
Look through a batch of .zip files produced by FASTQC to obtain summary information across a range of files
}
\details{
This will look in a number of .zip files, as produced by FASTQC,
and will extract the PASS/FAIL/WARN information across the key report sections.

\code{fqcSummary} will return the PASS/FAIL/WARN information for any number of FASTQC reports.

\code{fqcSummaryPlot} will plot the output from the function \code{fqcSummary} using default parameters.

\code{extractFqcSummary} is the low-level function for a single file and is primarily to be
called internally by the other two functions.
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}
\seealso{
fastqc
}

