% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGS2R.R
\docType{package}
\name{NGS2R}
\alias{NGS2R}
\alias{NGS2R-package}
\title{A package that provides access to some commonly used NGS data analysis tools}
\description{
Uses a series of wrappers and connections to access NGS data and tools
}

