% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHisat2Log.R
\name{readHisat2Log}
\alias{readHisat2Log}
\title{Read a single log file as written by hisat2}
\usage{
readHisat2Log(file)
}
\arguments{
\item{file}{The path to a single file}
}
\value{
A \code{\link{data_frame}} or \code{\link{tbl_df}}
}
\description{
Takes a log file as created by hisat2, and loads as a data_frame with a simgle row
}
\details{
Loads in the log file and extracts the key values to form a summary.
Multiple log Files can be loaded using \code{lapply} followed by \code{bind_rows}
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}

