% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sam2bam.R
\name{sam2bam}
\alias{sam2bam}
\title{Convert SAM files to BAM files}
\usage{
sam2bam(inFiles, outFiles, cl = NULL, sort = TRUE, sortArgs = c(),
  sortPre = "Sorted", rmUnsort = TRUE, index = TRUE,
  exec = "/usr/bin/samtools")
}
\arguments{
\item{inFiles}{A character vector with the full path of the SAM file(s)}

\item{outFiles}{The names to give the output files.
If not specified, this will default to the same names as the input, but with the suffix .bam}

\item{cl}{A cluster as defined by the package \code{parallel}. If specified, the conversion will be run in parallel}

\item{exec}{The path to the samtools executable. Defaults to "/usr/bin/samtools".
This may vary depending on where you have installed \code{samtools}}
}
\value{
a list with the components \code{$inFiles} & \code{$outFiles}
}
\description{
Converts SAM files to BAM files using \code{samtools view}
}
\details{
This is a wrapper to the function \code{samtools view} set to the conversion of sam files to bam files.

Files will be converted and the names of the newly created bam files will be supplied in the output in the component \code{$outFiles}.

This function is designed for simple operations on files existing on a local HDD.
For more complex analyses or processes, please see the package \code{Rsamtools} on
Bioconductor (\url{http://bioconductor.org/packages/release/bioc/html/Rsamtools.html}).
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}

