% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samtoolsCount.R
\name{samtoolsCount}
\alias{samtoolsCount}
\title{Count the requested number of reads in a given BAM/SAM File}
\usage{
samtoolsCount(file, include = NULL, exclude = NULL, minQ = NULL,
  unique = FALSE, exec = "/usr/bin/samtools")
}
\arguments{
\item{file}{A single BAM/SAM file to query}

\item{include}{Sets the flags for inclusion. This is the command-line parameter "-f"}

\item{exclude}{Sets the flags for exclusion. This is the command-line parameter "-F"}

\item{minQ}{Sets the minimum mapping quality to include}

\item{unique}{Provide the number of unique reads which match the specified criteria. Defaults to FALSE}

\item{exec}{The path to the samtools executable. Defaults to "/usr/bin/samtools".
This may vary depending on where you have installed \code{samtools}}
}
\value{
an integer corresponding to the count of reads matching the specified filter
}
\description{
Uses the requested filters to provide the count of reads which match within the specified BAM or SAM file.
}
\details{
This is a wrapper to the function \code{samtools view} with the option \code{-c} set.
Any combination of flags or quality filters can be used to extract the requested counts.

This function is designed for simple operations on files existing on a local HDD.
For more complex analyses or processes, please see the package \code{Rsamtools} on
Bioconductor (\url{http://bioconductor.org/packages/release/bioc/html/Rsamtools.html}).
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}

