% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/samtoolsIndex.R
\name{samtoolsIndex}
\alias{samtoolsIndex}
\title{Index BAM files}
\usage{
samtoolsIndex(bamFiles, cl = NULL, exec = "/usr/bin/samtools")
}
\arguments{
\item{bamFiles}{A character vector with the full path of the BAM file(s)}

\item{cl}{A cluster as defined by the package \code{parallel}. If specified, the conversion will be run in parallel}

\item{exec}{The path to the samtools executable. Defaults to "/usr/bin/samtools".
This may vary depending on where you have installed \code{samtools}}
}
\value{
a list with the components \code{$bamFiles} & \code{$indexFiles}
}
\description{
Indexes BAM files using \code{samtools index}
}
\details{
This is a wrapper to the function \code{samtools index} set to create the index files (.bai) to a supplied vector of bam files.

#' This function is designed for simple operations on files existing on a local HDD.
For more complex analyses or processes, please see the package \code{Rsamtools} on
Bioconductor (\url{http://bioconductor.org/packages/release/bioc/html/Rsamtools.html}).
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}

