% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/splitBamUniq.R
\name{splitBamUniq}
\alias{splitBamUniq}
\title{Split a BAM file into unique alignments & multiple alignments}
\usage{
splitBamUniq(inFile, outPath, exec = "/usr/bin/samtools")
}
\arguments{
\item{inFile}{The BAM file which is to be split}

\item{outPath}{The directory to write to. Defaults to the same directory as \code{inFile}. Will be created if it doesn't already exist.}

\item{exec}{The path to the samtooles executable}
}
\value{
A list with the components \code{inFile} & \code{outFiles}
}
\description{
Creates two new BAM files based on a single BAM file
}
\details{
This is a wrapper to the function \code{samtools view} and \code{grep}.

Files will be split and the names of the newly created bam files will be supplied in the output in the component \code{$outFiles}.
Currently the conversion takes place via SAM files, which will be converted back to BAM files.
The SAM files will then be deleted.

This function is designed for simple operations on files existing on a local HDD.
For more complex analyses or processes, please see the package \code{Rsamtools} on
Bioconductor (\url{http://bioconductor.org/packages/release/bioc/html/Rsamtools.html})
}
\author{
Steve Pederson
}

