% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/STAR.R
\name{buildSTARIndex}
\alias{alignSTAR}
\alias{buildSTARIndex}
\title{A wrapper for the RNA-Seq Aligner STAR}
\usage{
buildSTARIndex(genome, indexPath, threads, gtfFile, overhang = 84,
  exec = "/usr/local/bin/STAR")

alignSTAR(inFiles, indexPath, outDir, outType = "BAM", sorted = TRUE,
  threads, prefix, args = c(), exec = "/usr/local/bin/STAR")
}
\arguments{
\item{genome}{The reference genome as a single fasta file.}

\item{indexPath}{The directory to write the index files to when bilding, or to look for the index files when aligning}

\item{threads}{The number of threads to utilise. Defults to half the available cores.}

\item{gtfFile}{The annotation file in \code{.gtf} format}

\item{overhang}{The length of the overhang permitted. This should be set as max(ReadLength) - 1. Will default to 84, but this may not be suitable so please check this parameter}

\item{exec}{The path to the STAR executable}

\item{inFiles}{The input file(s). Should be a single fastq file, or paired end reads from a single sample.}

\item{outDir}{The directory to write the alignments to.}

\item{outType}{Specify whether to write the output files as SAM or BAM format. Defaults to BAM}

\item{sorted}{Specify whether to sort the alignments (\code{TRUE}) in the output files. Defaults to true.}

\item{prefix}{Any prefix to add to the sam/bam files. Will default to the prefix on the fastq file(s)}

\item{args}{Any additional arguments to pass to STAR.
Please note these will not be checked for errors, so please check the STAR manual carefully \url{https://github.com/alexdobin/STAR}.}
}
\value{
All functions execute the tool on the command line, then return a \code{list} with key information
}
\description{
Builds a STAR index, or align to the genome using STAR.
}
\details{
\code{buildSTARIndex} will build the index for aligning using the STAR RNA-seq aligner.
This tool is built for multi-threading and will run much faster if this capbility is taken advantage of.

\code{alignSTAR} is the wrapper for aligning a single fastq file, or a single set of paired end reads.
For multiple samples, please use the additional wrappers \code{lapply} or \code{snow::clusterApply}.
However, if aligning in parallel, please set the parameter \code{threads=1} as the function will automatically use half of the available cores, as STAR itself is written wil mutli-threading capabilities.

The commands for correctly utilising the shared memory capabilities of STAR have been problematic to implement,
so if you specifically require these, please use STAR on the command line.
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}

